[ADD] Message: Added a promised interface to handle callback.
