+++ CHANGELOG.zh-CN.md
@@ -16,6 +16,7 @@ timeline: true
 ---
 
 - 🌟 `Notification`: 新增 foldCount 全局配置属性。
+- 🌟 `Message`: 增加 promise 化的回调接口。
 - 💄 `Step`: 优化 Steps 组件的导航条样式。
 - 🐞 `<pro>Table`: 修复动态筛选条初始化请求及保存传参问题。
