+++ components-pro/performance-table/Cell.tsx
@@ -8,7 +8,52 @@ import { LAYER_WIDTH } from './constants';
 import { isNullOrUndefined, defaultClassPrefix, getUnhandledProps, prefix } from './utils';
 import TableContext from './TableContext';
 import Column from './Column';
-import { CellProps } from './Cell.d';
+import { StandardProps, RowDataType } from './common';
+
+export interface CellProps extends StandardProps {
+  align?: 'left' | 'center' | 'right';
+  verticalAlign?: 'top' | 'middle' | 'bottom';
+  className?: string;
+  classPrefix?: string;
+  dataKey?: string;
+  isHeaderCell?: boolean;
+
+  width?: number;
+  height?: number | ((rowData: object) => number);
+  left?: number;
+  headerHeight?: number;
+
+  style?: React.CSSProperties;
+  firstColumn?: boolean;
+  lastColumn?: boolean;
+  hasChildren?: boolean;
+  children?: React.ReactNode | ((rowData: RowDataType, rowIndex: number) => React.ReactNode);
+
+  rowKey?: string | number;
+  rowIndex?: number;
+  rowSpan?: number;
+  rowData?: RowDataType;
+  depth?: number; // change `Cell` width and depth prop to optional
+
+  onTreeToggle?: (
+    rowKey?: string | number,
+    rowIndex?: number,
+    rowData?: RowDataType,
+    event?: React.MouseEvent,
+  ) => void;
+
+  renderTreeToggle?: (
+    expandButton: React.ReactNode,
+    rowData?: RowDataType,
+    expanded?: boolean,
+  ) => React.ReactNode;
+  renderCell?: (contentChildren: any) => React.ReactNode;
+  wordWrap?: boolean;
+  hidden?: boolean;
+  treeCol?: boolean;
+  expanded?: boolean;
+  isDragging?: boolean;
+}
 
 export const propTypes = {
   align: PropTypes.oneOf(['left', 'center', 'right']),
@@ -153,7 +198,7 @@ class Cell extends React.PureComponent<CellProps> {
       if (!isHeaderCell && rowData && cellRowSpan === 0 && fixed) {
         runInAction(() => {
           tableStore.rowZIndex.push(rowIndex);
-        })
+        });
       }
       return null;
     }
@@ -189,7 +234,7 @@ class Cell extends React.PureComponent<CellProps> {
       width,
       height: nextHeight,
       textAlign: align,
-      [rtl ? 'paddingRight' : 'paddingLeft']: this.isTreeCol() ? depth! * LAYER_WIDTH + 10 : null
+      [rtl ? 'paddingRight' : 'paddingLeft']: this.isTreeCol() ? depth! * LAYER_WIDTH + 10 : null,
     };
 
     if (verticalAlign) {
