+++ components-pro/performance-table/CellGroup.tsx
@@ -2,9 +2,21 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
+import { DraggableProvided, DraggableStateSnapshot } from 'react-beautiful-dnd';
 import { defaultClassPrefix, getUnhandledProps, prefix } from './utils';
-import { CellGroupProps } from './CellGroup.d';
 import TableContext from './TableContext';
+
+export interface CellGroupProps {
+  fixed?: 'left' | 'right';
+  width?: number;
+  height?: number;
+  left?: number;
+  style?: React.CSSProperties;
+  className?: string;
+  classPrefix?: string;
+  snapshot?: DraggableStateSnapshot,
+  provided?: DraggableProvided,
+}
 
 const propTypes = {
   fixed: PropTypes.oneOf(['left', 'right']),
