+++ components-pro/performance-table/Column.tsx
@@ -1,5 +1,37 @@
+import React, { ReactNode } from 'react';
 import PropTypes from 'prop-types';
-import { ColumnProps } from './Column.d';
+
+export interface ColumnProps {
+  align?: 'left' | 'center' | 'right';
+  verticalAlign?: 'top' | 'middle' | 'bottom';
+  width?: number;
+  fixed?: boolean | 'left' | 'right';
+  resizable?: boolean;
+  sortable?: boolean;
+  flexGrow?: number;
+  minWidth?: number;
+  colSpan?: number;
+  treeCol?: boolean;
+  hidden?: boolean;
+  titleEditable?: boolean;
+  /**
+   * 是否可隐藏，设为false时不会出现在列过滤选项中
+   * @default true
+   */
+  hideable?: boolean;
+  onResize?: (columnWidth?: number, dataIndex?: string) => void;
+  render?: (props: { rowData: any, rowIndex: number, dataIndex?: string | undefined }) => React.ReactNode;
+  dataIndex?: string;
+  key?: string;
+  sort?: number;
+  title?: React.ReactNode | (() => React.ReactNode);
+  /**
+   * column group header
+   */
+  header?: ReactNode;
+  children?: React.ReactElement<ColumnProps>[] | ColumnProps[];
+  type?: 'ColumnGroup';
+}
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 function Column(_props: ColumnProps) {
