--- components-pro/performance-table/ColumnResizeHandler.d.ts
@@ -1,25 +1,0 @@
-import * as React from 'react';
-export type FixedType = boolean | 'left' | 'right';
-export interface Client {
-  clientX?: number;
-  clientY?: number;
-  preventDefault?: Function;
-}
-
-export interface ColumnResizeHandlerProps {
-  height?: number;
-  defaultColumnWidth?: number;
-  columnLeft?: number;
-  columnFixed?: FixedType;
-  className?: string;
-  classPrefix?: string;
-  minWidth?: number;
-  style?: React.CSSProperties;
-  onColumnResizeStart?: (client: Client) => void;
-  onColumnResizeEnd?: (columnWidth?: number, cursorDelta?: number) => void;
-  onColumnResizeMove?: (columnWidth?: number, columnLeft?: number, columnFixed?: FixedType) => void;
-}
-
-declare const ColumnResizeHandler: React.ComponentType<ColumnResizeHandlerProps>;
-
-export default ColumnResizeHandler;
