+++ components-pro/performance-table/ColumnResizeHandler.tsx
@@ -7,8 +7,28 @@ import { DOMMouseMoveTracker } from 'dom-lib';
 import { defaultClassPrefix, getUnhandledProps } from './utils';
 import { addEvent, removeEvent } from './utils/domFns';
 import TableContext from './TableContext';
-import { ColumnResizeHandlerProps } from './ColumnResizeHandler.d';
 import { RESIZE_MIN_WIDTH } from './constants';
+
+export type FixedType = boolean | 'left' | 'right';
+export interface Client {
+  clientX?: number;
+  clientY?: number;
+  preventDefault?: Function;
+}
+
+export interface ColumnResizeHandlerProps {
+  height?: number;
+  defaultColumnWidth?: number;
+  columnLeft?: number;
+  columnFixed?: FixedType;
+  className?: string;
+  classPrefix?: string;
+  minWidth?: number;
+  style?: React.CSSProperties;
+  onColumnResizeStart?: (client: Client) => void;
+  onColumnResizeEnd?: (columnWidth?: number, cursorDelta?: number) => void;
+  onColumnResizeMove?: (columnWidth?: number, columnLeft?: number, columnFixed?: FixedType) => void;
+}
 
 const propTypes = {
   height: PropTypes.number,
