+++ components-pro/performance-table/HeaderCell.tsx
@@ -2,10 +2,32 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import ColumnResizeHandler from './ColumnResizeHandler';
-import { isNullOrUndefined, getUnhandledProps, defaultClassPrefix, prefix } from './utils';
-import Cell from './Cell';
-import { HeaderCellProps } from './HeaderCell.d';
+import { defaultClassPrefix, getUnhandledProps, isNullOrUndefined, prefix } from './utils';
+import Cell, { CellProps } from './Cell';
 import TableContext from './TableContext';
+
+export interface HeaderCellProps extends CellProps {
+  index?: number;
+  minWidth?: number;
+  sortColumn?: string;
+  sortType?: 'desc' | 'asc';
+  sortable?: boolean;
+  style?: React.CSSProperties;
+  resizable?: boolean;
+  onColumnResizeStart?: (columnWidth?: number, left?: number, fixed?: boolean) => void;
+  onColumnResizeEnd?: (
+    columnWidth?: number,
+    cursorDelta?: number,
+    dataKey?: any,
+    index?: number,
+  ) => void;
+  onResize?: (columnWidth?: number, dataKey?: string) => void;
+  onColumnResizeMove?: (columnWidth?: number, columnLeft?: number, columnFixed?: boolean) => void;
+  onSortColumn?: (dataKey?: string) => void;
+  flexGrow?: number;
+  fixed?: boolean | 'left' | 'right';
+  dataIndex?: string;
+}
 
 interface HeaderCelltate {
   columnWidth?: number;
