+++ components-pro/performance-table/Row.tsx
@@ -2,9 +2,22 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
+import { DraggableProvided, DraggableStateSnapshot } from 'react-beautiful-dnd';
 import { defaultClassPrefix, getUnhandledProps, prefix } from './utils';
 import TableContext from './TableContext';
-import { RowProps } from './Row.d';
+import { StandardProps } from './common';
+
+export interface RowProps extends StandardProps {
+  width?: number;
+  height?: number;
+  headerHeight?: number;
+  top?: number;
+  isHeaderRow?: boolean;
+  rowDraggable?: boolean;
+  rowRef?: React.Ref<any>;
+  provided?: DraggableProvided;
+  snapshot?: DraggableStateSnapshot;
+}
 
 const propTypes = {
   width: PropTypes.number,
