--- components-pro/performance-table/global.d.ts
@@ -1,4 +1,0 @@
-declare const __RSUITE_CLASSNAME_PREFIX__: boolean;
-declare const __DEV__: boolean;
-
-declare module 'dom-lib';
