--- components-pro/performance-table/index.d.ts
@@ -1,5 +1,0 @@
-export { default as Table } from './Table';
-export { default as Column } from './Column';
-export { default as Cell } from './Cell';
-export { default as HeaderCell } from './HeaderCell';
-export { default as ColumnGroup } from './ColumnGroup';
