+++ components-pro/table/style/index.tsx
@@ -1,6 +1,7 @@
 import 'choerodon-ui/lib/style/index.less';
 import 'choerodon-ui/lib/collapse/style'
 import 'choerodon-ui/lib/grid/style';
+import 'choerodon-ui/lib/tag/style';
 import './index.less';
 
 // style dependencies
@@ -28,3 +29,4 @@ import '../../tooltip/style';
 import '../../validator/style';
 import '../../switch/style';
 import '../../menu/style';
+import '../../dropdown/style';
