+++ components/message/demo/placement.md
@@ -1,5 +1,5 @@
 ---
-order: 3
+order: 4
 title:
   zh-CN: 消息位置
   en-US: Message Placement
