+++ components/message/style/index.tsx
@@ -1,2 +1,4 @@
 import '../../style/index.less';
 import './index.less';
+
+import '../../progress/style';
